# Move automatic credit strategy in the handler thread